test-eptri: use address 0 for transfers
Due to a bug in iverilog, writes to wishbone are ignored towards the
start of the simulation. The signals can be seen, but for whatever
reason the registers simply are not updated.

As the simulation progresses, this limitation goes away.

The `test_control_transfer_in_out` test starts out by setting
`usb_address` to `20`, and proceeds to communicate with the device as if
this were its address.

However, due to this iverilog bug, this address is ignored and the
device address remains at 0.

This is not an issue so long as `test_control_transfer_in_out` is not
run in isolation, and is not the first test to be run. However, if this
test is run by itself, it will fail.

Leave the host address as 0. This ensures that it can pass even when run
in isolation.

Signed-off-by: Sean Cross <sean@xobs.io>