test-eptri: work around iverilog CSR reset issue
With iverilog, for some reason the first few writes to Wishbone are
ignored.  This results in writes to `usb_address` being ignored.

If the first test that's run after reset sets the USB address to
something other than 0, then that write will be ignored.

This issue doesn't crop up if the test isn't the first one to be run,
even if `reset()` is called beforehand.

Signed-off-by: Sean Cross <sean@xobs.io>